Remove mount_compendium method, move engine routes to config/routes.rb
